Resolve new tangles
- `MessagePublishingErrorHandler <-> IntegrationContextUtils`
