+++ spring-integration-core/src/main/java/org/springframework/integration/channel/interceptor/WireTap.java
@@ -23,8 +23,8 @@ import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
 import org.springframework.context.Lifecycle;
-import org.springframework.integration.context.IntegrationContextUtils;
 import org.springframework.integration.core.MessageSelector;
+import org.springframework.integration.support.channel.ChannelResolverUtils;
 import org.springframework.jmx.export.annotation.ManagedAttribute;
 import org.springframework.jmx.export.annotation.ManagedOperation;
 import org.springframework.jmx.export.annotation.ManagedResource;
@@ -182,7 +182,7 @@ public class WireTap implements ChannelInterceptor, Lifecycle, VetoCapableInterc
 		String channelNameToUse = this.channelName;
 		if (channelNameToUse != null) {
 			this.channel =
-					IntegrationContextUtils.getChannelResolver(this.beanFactory)
+					ChannelResolverUtils.getChannelResolver(this.beanFactory)
 							.resolveDestination(channelNameToUse);
 			this.channelName = null;
 		}
