+++ spring-integration-core/src/main/java/org/springframework/integration/context/IntegrationObjectSupport.java
@@ -37,6 +37,7 @@ import org.springframework.expression.ExpressionParser;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
 import org.springframework.integration.support.DefaultMessageBuilderFactory;
 import org.springframework.integration.support.MessageBuilderFactory;
+import org.springframework.integration.support.channel.ChannelResolverUtils;
 import org.springframework.integration.support.context.NamedComponent;
 import org.springframework.integration.support.utils.IntegrationUtils;
 import org.springframework.lang.Nullable;
@@ -227,7 +228,7 @@ public abstract class IntegrationObjectSupport implements BeanNameAware, NamedCo
 
 	protected DestinationResolver<MessageChannel> getChannelResolver() {
 		if (this.channelResolver == null) {
-			this.channelResolver = IntegrationContextUtils.getChannelResolver(this.beanFactory);
+			this.channelResolver = ChannelResolverUtils.getChannelResolver(this.beanFactory);
 		}
 		return this.channelResolver;
 	}
